Mac OS X: Link iconv in libhivex
hivexml on OS X was failing with a symbol-not-found error while
dynamically linking.  Adding iconv to libhivex fixes the issue and lets
OS X process 'images/large' with hivexml.

It took careful iconv autotool additions to get compilation working in
Fedora as well, but these two lines build and run on OS X 10.6 and
10.8, and Fedora 17 and 18.

Unfortunately, a separate issue is revealed here in OS X, as some of
the hivexml value data appears as Kanji.  However, that is an issue
distinct from getting the program to run.

Signed-off-by: Alex Nelson <a.nelson@prometheuscomputing.com>